+++ src/app.tsx
@@ -46,8 +46,9 @@ import {
   TurnServerOptions,
 } from "./options";
 import { MetaProvider, Style } from "@solidjs/meta";
-import { Input } from "./components/ui/input";
 import { produce } from "solid-js/store";
+import { createQRCodeDialog } from "./components/create-qrcode-dialog";
+import { createForwardDialog } from "./components/forward-dialog";
 let wakeLock: WakeLockSentinel | null = null;
 const requestWakeLock = async () => {
   if (!navigator.wakeLock) {
@@ -65,94 +66,6 @@ const requestWakeLock = async () => {
     if (err instanceof Error)
       console.error(`${err.name}, ${err.message}`);
   }
-};
-
-const createQRCodeDialog = () => {
-  const { colorMode } = useColorMode();
-  const { open, Component: QRCodeDialogComponent } =
-    createDialog({
-      title: () => t("common.scan_qrcode_dialog.title"),
-      description: () => (
-        <>
-          <span class="text-lg font-bold">
-            {clientProfile.name}&nbsp;
-          </span>
-          <span class="text-sm text-muted-foreground">
-            {t("common.scan_qrcode_dialog.invite", {
-              room: clientProfile.roomId,
-            })}
-          </span>
-        </>
-      ),
-      content: () => {
-        const url = joinUrl();
-        return (
-          <div class="flex select-none flex-col items-center gap-2">
-            <div
-              onContextMenu={(e) => {
-                e.preventDefault();
-                navigator.clipboard
-                  .writeText(url)
-                  .then(() => {
-                    toast.success(
-                      t(
-                        "common.notification.link_copy_success",
-                      ),
-                    );
-                  })
-                  .catch(() => {
-                    toast.error(
-                      t("common.notification.copy_failed"),
-                    );
-                  });
-              }}
-            >
-              <QRCode
-                value={url}
-                dark={
-                  colorMode() === "dark"
-                    ? "#ffffff"
-                    : "#000000"
-                }
-                light="#00000000"
-                logo={clientProfile.avatar ?? undefined}
-                logoShape="circle"
-              />
-            </div>
-            <Input
-              class="h-8 w-full max-w-sm select-all whitespace-pre-wrap break-all
-                text-xs hover:underline"
-              readOnly
-              onContextMenu={async (e) => {
-                e.preventDefault();
-                navigator.clipboard
-                  .writeText(url)
-                  .then(() => {
-                    toast.success(
-                      t(
-                        "common.notification.link_copy_success",
-                      ),
-                    );
-                  })
-                  .catch(() => {
-                    toast.error(
-                      t("common.notification.copy_failed"),
-                    );
-                  });
-              }}
-              value={joinUrl()}
-            />
-            <p>
-              {t("common.scan_qrcode_dialog.description")}
-            </p>
-            <p class="mt-2 text-sm text-muted-foreground">
-              {t("common.scan_qrcode_dialog.tip")}
-            </p>
-          </div>
-        );
-      },
-    });
-  return { open, Component: QRCodeDialogComponent };
 };
 
 const InnerApp = (props: ParentProps) => {
@@ -278,11 +191,37 @@ const InnerApp = (props: ParentProps) => {
   onCleanup(() => {
     wakeLock?.release();
   });
+
+  const {
+    forwardTarget: shareTarget,
+    Component: ForwardDialogComponent,
+  } = createForwardDialog();
+
+  onMount(() => {
+    const onMessage = (ev: MessageEvent) => {
+      if (ev.data.action === "share-target") {
+        window.focus();
+        shareTarget(ev.data.data as ShareData);
+      }
+    };
+    navigator.serviceWorker.addEventListener(
+      "message",
+      onMessage,
+    );
+    onCleanup(() => {
+      navigator.serviceWorker.removeEventListener(
+        "message",
+        onMessage,
+      );
+    });
+  });
+
   return (
     <>
       <RoomDialogComponent />
       <QRCodeDialogComponent />
       <AboutDialogComponent />
+      <ForwardDialogComponent />
       <div
         class="sticky top-0 z-50 flex h-12 w-full flex-wrap items-center
           gap-4 border-b border-border bg-background/80 px-2