+++ src/components/ui/checkbox.tsx
@@ -37,21 +37,22 @@ export const CheckboxControl = <
       />
       <CheckboxPrimitive.Control
         class={cn(
-          `h-4 w-4 shrink-0 rounded-sm border border-primary shadow
-          transition-shadow focus-visible:outline-none
-          focus-visible:ring-[1.5px] focus-visible:ring-ring
-          data-[disabled]:cursor-not-allowed data-[checked]:bg-primary
+          `size-4 shrink-0 rounded-sm border border-primary shadow
+          transition-shadow data-[disabled]:cursor-not-allowed
+          data-[checked]:bg-primary
           data-[checked]:text-primary-foreground
-          data-[disabled]:opacity-50`,
+          data-[disabled]:opacity-50 focus-visible:outline-none
+          focus-visible:ring-[1.5px] focus-visible:ring-ring`,
           local.class,
         )}
         {...rest}
       >
-        <CheckboxPrimitive.Indicator class="flex items-center justify-center text-current">
+        <CheckboxPrimitive.Indicator class="relative size-4">
           <svg
             xmlns="http://www.w3.org/2000/svg"
             viewBox="0 0 24 24"
-            class="h-4 w-4"
+            class="absolute left-1/2 top-1/2 size-4 -translate-x-1/2
+              -translate-y-1/2"
           >
             <path
               fill="none"
@@ -59,7 +60,7 @@ export const CheckboxControl = <
               stroke-linecap="round"
               stroke-linejoin="round"
               stroke-width="2"
-              d="m5 12l5 5L20 7"
+              d="m5 12l4 4 8-8"
             />
             <title>Checkbox</title>
           </svg>