+++ src/routes/client/redirect.tsx
@@ -1,0 +1,6 @@
+const Redirect = () => {
+ 
+  return <></>;
+};
+
+export default Redirect;